Work on git Issue #1 - Code cleanup EE and AD service

+++ gemma-mda/src/main/java/ubic/gemma/model/expression/arrayDesign/ArrayDesignService.java
@@ -118,7 +118,7 @@ public interface ArrayDesignService {
     Map<CompositeSequence, BioSequence> getBioSequences( ArrayDesign arrayDesign );
 
     @Secured({ "IS_AUTHENTICATED_ANONYMOUSLY", "ACL_SECURABLE_READ" })
-    Integer getCompositeSequenceCount( ArrayDesign arrayDesign );
+    Long getCompositeSequenceCount( ArrayDesign arrayDesign );
 
     @Secured({ "IS_AUTHENTICATED_ANONYMOUSLY", "ACL_SECURABLE_READ" })
     Collection<CompositeSequence> getCompositeSequences( ArrayDesign arrayDesign );
